Make attribute choices strings, not enums
Adding or removing an attribute could change the order of the enums,
which would break existing data. This also meantthat exporting
only worked if the correct entry was actually selected, since this
determined the set of items that were in the enum.

Now, the set of attributes is stored in a collection property in
the scene, and it gets updated whenever the package or entry is
switched. There is also a manual refresh button, which should only be
needed immediately after adding a new attribute.