Store vertices directly, not with type punning
The file version has been incremented to 2.

Floats can count up to around 16 million, which is plenty. This allows
for vertex IDs to be added by a modifier, rather than being written
into the UV map directly.

This is necessary if the model is being procedurally generated, since
the vertices don't even exist in the base object data.

Since a modifier won't work for meshes with shape keys, there is still
an option to bake the data directly into the UV map.

Also, the exporter now checks if there are enough UV maps present and
creates new ones if necessary.

A .blend library has been added. It contains the geometry node tree
used to store vertex IDs.